htlcswitch/link: consolidate ExpiryTooSoon checks
This commit restructures the checks done on
the exit hop's CLTV delta, such that we only
apply the check once. Currently we make two
distinct checks, one for the grace delta, and
one for the invoice's required CLTV delta.

We resolve this by taking the maximum of the two
to use our as our expected height, as this is the
stricter check that was being asserted with
two separate cases.