channeldb/cache: add edgeCache and channelCache
This commit adds two simple caches:
 - edgeCache: caches channel existence and update timestamps
 - channelCache: caches chan info and chan updates

The edgeCache is used by the router to test for eviction
and staleness of incoming channels, which uses HasChannelEdge
underneath. When merged with the persistent zombie update
tracking, it will also be extended to store the zombie status.
The channelCache is used by the gossiper when responding to
ChanUpdatesInHorizon, by avoiding repetitive deserialization
of recently loaded infos and policies.

Both caches are implemented using a random eviction
policy, since the access patterns from our peers may
cause performance degradation under certain access patterns,
either accidental or maliciously crafted. For instance, it
is easy to craft an access pattern that foil an LRU cache
if the entire query size does not fit in memory, as a result
of the linear seek on disk. In practice, the randomized cache
quickly converges back to high hit rate, depending on the
work loads requested.